+++ src/main/java/com/example/web/controller/TeacherController.java
@@ -1,7 +1,5 @@
 package com.example.web.controller;
 
-import com.example.web.repository.QuestionAnswerTableRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.servlet.ModelAndView;
