+++ src/main/java/com/example/web/controller/TestController.java
@@ -1,6 +1,6 @@
 package com.example.web.controller;
 
-import com.example.web.repository.QuestionAnswerTableRepository;
+import com.example.web.dao.repository.QuestionAnswerTableRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
