+++ src/main/java/com/example/web/dao/repository/CourseTableRepository.java
@@ -1,7 +1,6 @@
-package com.example.web.repository;
+package com.example.web.dao.repository;
 
 import com.example.web.bean.CourseTable;
-import com.example.web.bean.CrmUserTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
