+++ src/main/java/com/example/web/dao/repository/HwForStudentRepository.java
@@ -1,9 +1,9 @@
-package com.example.web.repository;
+package com.example.web.dao.repository;
 
 import com.example.web.bean.HwFromStudentTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HwFromStudentRepository extends JpaRepository<HwFromStudentTable, Long> {
+public interface HwForStudentRepository extends JpaRepository<HwFromStudentTable, Long> {
 }
